# Conflicts:
#	CHANGES.txt
#	README.md

+++ tests/vagrant/preloaded/create_preloaded.sh
@@ -13,10 +13,37 @@ done
 
 vagrant box update
 
-for box in "preloaded-ubuntu-bionic64" "preloaded-ubuntu-focal64"
+for box in "preloaded-ubuntu-focal64" "preloaded-ubuntu-jammy64"
 do
-    vagrant up $box
+    vagrant up $box | tee /tmp/$box.out
     upcode=$?
+    
+    if [ $upcode -eq 0 -a ! -e "./prepcode.txt" ] && grep -F 'Authentication failure' /tmp/$box.out >/dev/null; then
+        # note: upcode is 0 only if config.vm.boot_timeout is set.
+        # If this works it may be an indication that ruby's internal
+        # ssh does not support the algorithm required by the server,
+        # or the public key does not match (vagrant and vm out of
+        # sync)
+        echo ""
+        echo "VAGRANT AUTHENTICATION FAILURE - TRYING LOOSER ALLOWED SSHD ALGS"
+        if vagrant ssh $box -c "sudo bash -c 'echo PubkeyAcceptedAlgorithms +ssh-rsa > /etc/ssh/sshd_config.d/ciab.conf; sudo systemctl restart sshd'"; then
+            vagrant halt $box
+            vagrant up $box
+            upcode=$?
+        fi
+    fi
+        
+    if [ $upcode -ne 0 -a ! -e "./prepcode.txt" ]
+    then
+        # a reboot may be necessary if guest addtions was newly
+        # compiled by vagrant plugin "vagrant-vbguest"
+        echo ""
+        echo "VAGRANT UP RETURNED $upcode -- RETRYING AFTER REBOOT"
+        vagrant halt $box
+        vagrant up $box
+        upcode=$?
+    fi
+
     prepcode=$(cat "./prepcode.txt")
     rm -f prepcode.txt
     echo ""
